refactor(neo4j-storage): extract mappings + read mixin (issue #50 sub 1/3)
Issue #50 (EPIC-08-ST-01), Sub-Slice 1/3 — first cut of the
Read/Write/Search split.

- New: backend/app/storage/neo4j_mappings.py (118 LOC)
  - node_to_dict, edge_to_dict: pure functions, drop the embedding /
    name_lower / fact_embedding internal fields, robust JSON parser
    for attributes_json
  - sanitize_label: ASCII identifier whitelist for LLM-emitted labels
    (moved from neo4j_storage module-level so both read and write
    paths can import it without a circular dependency)
- New: backend/app/storage/neo4j_read.py (381 LOC)
  - Neo4jReadMixin holds all 10 read methods: get_ontology,
    get_all_nodes, get_node, get_node_edges, get_nodes_by_label,
    get_filtered_entities_with_edges, get_all_edges, get_edges_at_round,
    get_graph_info, get_graph_data
  - Mixin pattern preserves shared state (self._driver,
    self._call_with_retry) through MRO — no constructor refactor
- Changed: backend/app/storage/neo4j_storage.py 1127 -> 701 LOC (-426,
  -37.8%)
  - class Neo4jStorage(Neo4jReadMixin, GraphStorage)
  - _node_to_dict / _edge_to_dict kept as staticmethod re-exports of
    the module functions so any external subclass keeps working
  - _LABEL_SAFE_RE + _sanitize_label module-level removed (moved to
    neo4j_mappings); import re drops out
- New: backend/tests/test_neo4j_mappings.py (23 tests, 3 classes)
  - node mapping (7): minimal, JSON parsing, broken JSON fallback,
    internal field stripping, label filtering, missing optionals
  - edge mapping (6): minimal with temporal fields, fact_embedding
    stripped, episode_ids scalar wrapping, default reinforced_count,
    broken JSON
  - sanitize_label (10): ASCII pass, underscore, whitespace, umlaute
    stripped, Entity/empty/non-string rejected, starting digit
    rejected, backtick injection neutralized, 50-char limit

Acceptance criteria:
- [x] "Mappings sind zentral wiederverwendbar" — covered by this slice
- [ ] "Schreiblogik, Leselogik und Suche liegen nicht mehr in einer
  Datei" — read path covered, write (sub-slice 2) and search
  (sub-slice 3) follow

npm run check: 5/5 green (671 backend +23, 40 frontend, build ok).
Existing test_neo4j_resilience, test_neo4j_filtered_entities,
test_neo4j_ontology_wiring (18 tests) stay green — behaviour identical
through Mixin MRO.

Refs #50
